requirements.txt
keras
numpy
tensorflow>=2.2.0
opencv-python
pillow
pandas
pylint
sklearn
